LayoutTests/imported/w3c/web-platform-tests/content-security-policy/navigate-to/meta-refresh-blocked.sub-expected.txt



FAIL Test that the child iframe navigation is not allowed assert_equals: expected "fail" but got "success"
FAIL Violation report status OK. assert_true: No CSP report sent, but expecting one. expected true got false
